requirements.txt
awscli==1.10.12
boto==2.39.0
boto3==1.2.6
coverage==4.1b2
Fabric==1.10.2
Flask==0.10.1
ipython==4.1.2
pylint==1.5.4
python-slugify==1.2.0
PyYAML==3.11
requests==2.9.1
troposphere==1.5.0
git+https://github.com/nose-devs/nose@master
